+++ app/Controllers/CDR/Groups.php
@@ -6,14 +6,21 @@ use App\Controllers\BaseController;
 class Groups extends BaseController
 {
 
-	public function __construct()
-	{
+	public function __construct(){
 		$this->authorize = service('authorization');
 		helper('array');
 	}
-
-	public function list()
-	{
+	public function add(){
+		$data = [
+			'menuActive' => [
+				'active' => 'config',
+				'col' => 'groups',				
+			],
+			'groups' => $this->authorize->groups(),
+		];
+		return view('config/groups/add',array_merge($data, $this->data));
+	}
+	public function list(){
 		$data = [
 			'scripts' => [
 				'DataTables' => 'assets/plugins/data-tables/jquery.datatables.min.js',
@@ -73,18 +80,26 @@ class Groups extends BaseController
 		if (! $this->validate($rules))
 		{
 			return redirect()->back()->withInput()->with('errors', $this->validator->getErrors());
-		}
-		if(!$groupId = $this->authorize->group($this->request->getPost('id'))){
-			return redirect()->to('config/groups')->with('message', 'Esse grupo não existe');
-		}
+		}		
 		$name = $this->request->getPost('name');
 		$description = $this->request->getPost('description');
-
-		if(!$this->authorize->updateGroup($groupId->id, $name, $description)){
-			return redirect()->to('config/groups')->with('message', 'Ocorreu um erro ao alterar o nome e descrição');
+		if ( is_null($this->request->getPost('id')) ){
+			if( !$id = $this->authorize->createGroup($name, $description)){
+				return redirect()->to('config/groups')->with('message', 'Ocorreu um erro ao criar o grupo');
+			}
+			$mensagem = 'Criado com sucesso';
+		}else{
+			$id = $this->request->getPost('id');
+			if(!$this->authorize->group($id)){
+				return redirect()->to('config/groups')->with('message', 'Esse grupo não existe');
+			}
+			if(!$this->authorize->updateGroup($id, $name, $description)){
+				return redirect()->to('config/groups')->with('message', 'Ocorreu um erro ao alterar o nome e descrição');
+			}
+			$mensagem = 'Atualizado com sucesso';
 		}
 
-		$permitions = array_column($this->authorize->groupPermissions($groupId->id), 'name');
+		$permitions = array_column($this->authorize->groupPermissions($id), 'name');
 		$usersAdd = $this->request->getPost('users-add');
 		$usersEdit = $this->request->getPost('users-edit');
 		$usersList = $this->request->getPost('users-list');
@@ -92,30 +107,30 @@ class Groups extends BaseController
 		$usersStatus = $this->request->getPost('users-status');	
 
 		if($usersAdd && !in_array('users-add',$permitions)){
-			$this->authorize->addPermissionToGroup('users-add',$groupId->id);
+			$this->authorize->addPermissionToGroup('users-add',$id);
 		}else if (in_array('users-add',$permitions) && !$usersAdd ){
-			$this->authorize->removePermissionFromGroup('users-add',$groupId->id);
+			$this->authorize->removePermissionFromGroup('users-add',$id);
 		}
 		if($usersEdit && !in_array('users-edit',$permitions)){
-			$this->authorize->addPermissionToGroup('users-edit',$groupId->id);
+			$this->authorize->addPermissionToGroup('users-edit',$id);
 		}else if (in_array('users-edit',$permitions) && !$usersEdit ){
-			$this->authorize->removePermissionFromGroup('users-edit',$groupId->id);
+			$this->authorize->removePermissionFromGroup('users-edit',$id);
 		}
 		if($usersList && !in_array('users-list',$permitions)){
-			$this->authorize->addPermissionToGroup('users-list',$groupId->id);
+			$this->authorize->addPermissionToGroup('users-list',$id);
 		}else if (in_array('users-list',$permitions) && !$usersList ){
-			$this->authorize->removePermissionFromGroup('users-list',$groupId->id);
+			$this->authorize->removePermissionFromGroup('users-list',$id);
 		}
 		if($usersRemove && !in_array('users-remove',$permitions)){
-			$this->authorize->addPermissionToGroup('users-remove',$groupId->id);
+			$this->authorize->addPermissionToGroup('users-remove',$id);
 		}else if (in_array('users-remove',$permitions) && !$usersRemove ){
-			$this->authorize->removePermissionFromGroup('users-remove',$groupId->id);
+			$this->authorize->removePermissionFromGroup('users-remove',$id);
 		}
 		if($usersStatus && !in_array('users-status',$permitions)){
-			$this->authorize->addPermissionToGroup('users-status',$groupId->id);
+			$this->authorize->addPermissionToGroup('users-status',$id);
 		}else if (in_array('users-status',$permitions) && !$usersStatus ){
-			$this->authorize->removePermissionFromGroup('users-status',$groupId->id);
+			$this->authorize->removePermissionFromGroup('users-status',$id);
 		}
-		return redirect()->back();
+		return redirect()->to('config/groups')->with('message', $mensagem);
 	}
 }